requirements.txt
diskcache==5.6.3
Jinja2==3.1.6
llama_cpp_python==0.3.16
MarkupSafe==3.0.3
numpy==2.4.1
packaging==26.0
setuptools==80.10.2
typing_extensions==4.15.0
wheel==0.46.3
